fix/#64-중간 점검
