+++ hearo/hearo/Sources/Presentations/Finish/ViewModel/FinishViewModel.swift
@@ -19,6 +19,7 @@ class FinishViewModel: ObservableObject {
     private func goToHomeWithDelay() {
         
         DispatchQueue.main.asyncAfter(deadline: .now() + 2) { [weak self] in
+            triggerSuccessHaptic()
             self?.appRootManager.currentRoot = .home
         }
     }
