--- hearo/hearo/Sources/Presentations/Working/ViewModel/HornSoundClassifier.mlproj/Project.json
@@ -1,1 +1,0 @@
-{"version":"1.0","authorName":"규북","title":"","license":"","location":"","modifiedDate":"2024-10-12T14:29:05Z","id":"54841784-ee7e-4716-9279-0ca3e1df9e7e","createdDate":"2024-10-12T14:29:05Z","description":"","projectName":""}
